Se refactoriza chat y se divide en servicio que maneja logica central de entrevista y cache entrevista
