+++ src/app/components/zona-entrevista/multipasos/multipasos.module.ts
@@ -11,7 +11,6 @@ import { ProgressBarModule } from 'primeng/progressbar';
 import { ToastModule } from 'primeng/toast';
 import { TooltipModule } from 'primeng/tooltip';
 import { Paso2Component } from './paso-2/paso-2.component';
-import { Paso3Component } from './paso-3/paso-3.component';
 import { Paso4Component } from './paso-4/paso-4.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { StepsModule } from 'primeng/steps';
@@ -20,23 +19,17 @@ import { InputTextModule } from 'primeng/inputtext';
 import { StepperModule } from 'primeng/stepper';
 import { DividerModule } from 'primeng/divider';
 import { ZonaEntrevistaRoutingModule } from '../zona-entrevista-routing.module';
-import { ChatAppComponent } from './paso-3/chat/chat-app/chat-app.component';
-import { UsersChatsComponent } from './paso-3/chat/chat-app/users-chats/users-chats.component';
-import { VideoChatComponent } from './paso-3/chat/video-chat/video-chat.component';
-import { ChatMenuComponent } from './paso-3/chat/chat-menu/chat-menu.component';
 import { WaitingAreaComponent } from './waiting-area/waiting-area.component';
 import { CarouselModule } from 'ngx-owl-carousel-o';
 import { CollapseModule } from 'ngx-bootstrap/collapse';
+import { Paso3Module } from './paso-3/paso-3.module';
+import { Paso3Component } from './paso-3/paso-3.component';
 @NgModule({
   declarations: [
     MultipasosComponent,
     Paso2Component,
     Paso3Component,
     Paso4Component,
-    ChatAppComponent,
-    VideoChatComponent,
-    UsersChatsComponent,
-    ChatMenuComponent,
     WaitingAreaComponent,
   ],
 
@@ -60,6 +53,7 @@ import { CollapseModule } from 'ngx-bootstrap/collapse';
     DividerModule,
     CarouselModule,
     CollapseModule.forRoot(),
+    Paso3Module
   ],
 
   exports: [
