+++ src/app/components/zona-entrevista/multipasos/paso-3/base-entrevista/base-entrevista.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { BaseEntrevistaComponent } from './base-entrevista.component';
+
+describe('BaseEntrevistaComponent', () => {
+  let component: BaseEntrevistaComponent;
+  let fixture: ComponentFixture<BaseEntrevistaComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [BaseEntrevistaComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(BaseEntrevistaComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
