+++ src/app/components/zona-entrevista/multipasos/paso-3/base-entrevista/entrevista-component-interface.ts
@@ -1,0 +1,6 @@
+export interface EntrevistaComponentInterface {
+    initializeInterview(): void;
+    handleUserResponse(response: string): void;
+    showNextQuestion(): void;
+    finalizeInterview(): void;
+  }
