--- src/app/components/zona-entrevista/multipasos/paso-3/chat/chat-app/chat-app.component.html
@@ -1,21 +1,0 @@
-<div class="container-fluid">
-   <div class="row">
-     <div class="col call-chat-sidebar">
-       <div class="card">
-         <div class="card-body chat-body">
-           <div class="chat-box">
-              <app-users-chats></app-users-chats>
-           </div>
-         </div>
-       </div>
-     </div>
-     <div class="col call-chat-body">
-       <div class="card">
-         <div class="card-body p-0">
-           <div class="row chat-box">
-           </div>
-         </div>
-       </div>
-     </div>
-   </div>
- </div>
