--- src/app/components/zona-entrevista/multipasos/paso-3/chat/chat-app/chat-app.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-@Component({
-  selector: 'app-chat-app',
-  templateUrl: './chat-app.component.html',
-  styleUrls: ['./chat-app.component.scss']
-})
-export class ChatAppComponent implements OnInit {
-  public clickedEvent: any;
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-  childEventClicked(user: Event) {
-    this.clickedEvent = user;
-  }
-}
