--- src/app/components/zona-entrevista/multipasos/paso-3/chat/chat-app/users-chats/users-chats.component.html
@@ -1,35 +1,0 @@
-<div class="chat-left-aside">
-   <div class="media"><img class="rounded-circle user-image" src="assets/images/user/12.png" alt="">
-     <div class="media-body">
-       <div class="about">
-         <div class="name f-w-600">{{profile.name}}</div>
-         <div class="status">{{profile.status}}</div>
-       </div>
-     </div>
-   </div>
-   <div class="people-list" id="people-list">
-     <div class="search">
-       <form class="theme-form">
-         <div class="form-group">
-           <input class="form-control" type="text" placeholder="search"><i class="fa fa-search"></i>
-         </div>
-       </form>
-     </div>
-     <ul class="list custom-scrollbar">
-      <ng-container *ngFor="let user of searchUsers">
-        <li class="clearfix">
-          <div class="media">
-            <img class="rounded-circle user-image" [src]="user.profile" alt="">
-            <div class="status-circle away" [ngClass]="{'online' : user.online, 'offline' : !user.online}"></div>
-            <div class="media-body">
-              <div class="about">
-                <div class="name">{{user.name}}</div>
-                <div class="status">{{user.status}}</div>
-              </div>
-            </div>
-          </div>
-        </li>
-      </ng-container>
-     </ul>
-   </div>
- </div>
