+++ src/app/components/zona-entrevista/multipasos/paso-3/chat/video-chat/video-chat.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { VideoChatComponent } from './video-chat.component';
+
+describe('VideoChatComponent', () => {
+  let component: VideoChatComponent;
+  let fixture: ComponentFixture<VideoChatComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [VideoChatComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(VideoChatComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
