+++ src/app/components/zona-entrevista/multipasos/paso-3/paso-3.component.html
@@ -1,7 +1,7 @@
 <p-toast></p-toast>
 
-<app-video-chat class="mt-3" [idEntrevista]="idEntrevista" (respuestasEntrevista)="submitAnswers($event)"></app-video-chat>
-
+<!--app-video-chat class="mt-3" [idEntrevista]="idEntrevista" (respuestasEntrevista)="submitAnswers($event)"></app-video-chat-->
+<app-text-chat class="mt-3" [idEntrevista]="idEntrevista" (respuestasEntrevista)="submitAnswers($event)"></app-text-chat>
 <p-dialog header="Header" [(visible)]="visible" [modal]="true" [breakpoints]="{ '960px': '75vw' }" [style]="{ width: '50vw' }" [draggable]="false" [resizable]="false">
     
   <ng-template pTemplate="header">
