+++ src/app/components/zona-entrevista/multipasos/waiting-area/waiting-area.component.html
@@ -4,7 +4,7 @@
             <div class="row prueba">
                 <div class="col-md-1"></div>
                 <div class="col-md-10 text-center">
-                    <img src="../../../../../assets/images/chatbot.png" width="180" alt="chatbot" srcset="">
+                    <img src="assets/images/chatbot.png" width="180" alt="chatbot" srcset="">
                     <h3 class="animate-charcter display-3" [id]="'typewriter-' + 1"><span class="cursor"><b>|</b></span></h3>
                     
                 </div>
