+++ src/environments/environment.prod.ts
@@ -1,18 +1,18 @@
 export const environment = {
   production: true,
-  // OAUTH MODULE -- DEL APP MODULE
-  urlBase: 'https://keycloak.kahop.com/',
+    // OAUTH MODULE -- DEL APP MODULE
+    urlBase: 'https://keycloak.pruebas-entrevistador-inteligente.site/',
 
-  // INTEGRADOR SERVICE
-  orquestadorURL: 'https://gw.kahop.com/api/orquestador/v1',
+    // INTEGRADOR SERVICE
+    orquestadorURL: 'https://gateway.pruebas-entrevistador-inteligente.site/api/orquestador/v1',
+  
+    // FEEDBACK SERVICE
+    feedbackURL: 'https://gateway.pruebas-entrevistador-inteligente.site/api/administrador-entrevista/v1',
+  
+    // FEEDBACK SERVICE
+    notifcacionesURL: 'https://gateway.pruebas-entrevistador-inteligente.site/api/notificaciones/v1',
 
-  // FEEDBACK SERVICE
-  feedbackURL: 'https://gw.kahop.com/api/administrador-entrevista/v1',
+    chatURL: 'https://gateway.pruebas-entrevistador-inteligente.site/api/chats/v1',
 
-  // FEEDBACK SERVICE
-  notifcacionesURL: 'https://gw.kahop.com/api/notificaciones/v1',
-
-  chatURL: 'https://gw.kahop.com/api/chats/v1',
-
-  landingApp: 'https://www.kahop.com/es/pagina-principal/',
+    landingApp: 'https://www.pruebas-entrevistador-inteligente.site/es/pagina-principal/',
 };
